+++ src/methods/crossmappings/estimators/Ensemble.jl
@@ -9,7 +9,7 @@ using the given estimator `est` (e.g. [`RandomVectors`](@ref))
 
 ## Examples
 
-- [Example 1](@ref example_ConvergentCrossMapping_reproducing_Sugihara_Fig3A): 
+- [Example 1](@ref example_ConvergentCrossMapping_reproducing_sugihara): 
     Reproducing Figure 3A from [Sugihara2012](@citet).
 """
 Base.@kwdef struct Ensemble{M <: CrossmapEstimator{<:CrossmapMeasure}} # todo: perhaps just use a more general Ensemble?