+++ src/methods/information/definitions/conditional_mutual_informations/CMIShannon.jl
@@ -51,7 +51,7 @@ differential entropies.
 
 - [Example 1](@ref example_CMIShannon_EntropyDecomposition_Kraskov): 
     [`EntropyDecomposition`](@ref) with [`Kraskov`](@ref) estimator.
-- [Example 2](@ref CMIShannon_EntropyDecomposition_ValueBinning):
+- [Example 2](@ref example_CMIShannon_EntropyDecomposition_ValueBinning):
     [`EntropyDecomposition`](@ref) with [`ValueBinning`](@ref) estimator.
 - [Example 3](@ref example_CMIShannon_MIDecomposition_KSG1): 
     [`MIDecomposition`](@ref) with [`KraskovStögbauerGrassberger1`](@ref) estimator.